renamed ExplorativeSubgraphIsomorphism
* renamed package from isomorphism to preserving
* renamed operators to ExplorativePatternMatching
* updated package infos
* fixes #349
