+++ gradoop-flink/src/main/java/org/gradoop/flink/model/impl/operators/matching/preserving/explorative/debug/PrintEdgeStep.java
@@ -15,34 +15,23 @@
  * along with Gradoop. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.gradoop.flink.model.impl.operators.matching.isomorphism.explorative.debug;
+package org.gradoop.flink.model.impl.operators.matching.preserving.explorative.debug;
 
-import org.apache.commons.lang3.StringUtils;
 import org.apache.log4j.Logger;
+import org.gradoop.flink.model.impl.operators.matching.preserving.explorative.tuples.EdgeStep;
 import org.gradoop.flink.model.impl.operators.matching.common.debug.Printer;
-import org.gradoop.flink.model.impl.operators.matching.isomorphism.explorative.tuples.EmbeddingWithTiePoint;
-
-import java.util.Arrays;
 
 /**
- * Debug output for {@link EmbeddingWithTiePoint}.
+ * Debug output for {@link EdgeStep}.
  *
  * @param <K> key type
  */
-public class PrintEmbeddingWithTiePoint<K>
-  extends Printer<EmbeddingWithTiePoint<K>, K> {
+public class PrintEdgeStep<K>
+  extends Printer<EdgeStep<K>, K> {
   /**
    * Logger
    */
-  private static final Logger LOG = Logger.getLogger(
-    PrintEmbeddingWithTiePoint.class);
-
-  /**
-   * Constructor
-   */
-  public PrintEmbeddingWithTiePoint() {
-    this(false, "");
-  }
+  private static final Logger LOG = Logger.getLogger(PrintEdgeStep.class);
 
   /**
    * Constructor
@@ -50,18 +39,16 @@ public class PrintEmbeddingWithTiePoint<K>
    * @param isIterative true, if used in iterative context
    * @param prefix      prefix for debug string
    */
-  public PrintEmbeddingWithTiePoint(boolean isIterative, String prefix) {
+  public PrintEdgeStep(boolean isIterative, String prefix) {
     super(isIterative, prefix);
   }
 
   @Override
-  protected String getDebugString(EmbeddingWithTiePoint<K> embedding) {
-    return String.format("(([%s],[%s]),%s)",
-      StringUtils.join(convertList(Arrays.asList(
-        embedding.getEmbedding().getVertexMappings()), true), ','),
-      StringUtils.join(convertList(Arrays.asList(
-        embedding.getEmbedding().getEdgeMappings()), false), ','),
-      vertexMap.get(embedding.getTiePointId()));
+  protected String getDebugString(EdgeStep<K> edgeStep) {
+    return String.format("(%s,%s,%s)",
+      edgeMap.get(edgeStep.getEdgeId()),
+      vertexMap.get(edgeStep.getTiePoint()),
+      vertexMap.get(edgeStep.getNextId()));
   }
 
   @Override
