+++ gradoop-flink/src/main/java/org/gradoop/flink/model/impl/operators/matching/preserving/explorative/debug/PrintEmbeddingWithTiePoint.java
@@ -15,24 +15,36 @@
  * along with Gradoop. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.gradoop.flink.model.impl.operators.matching.isomorphism.explorative.debug;
+package org.gradoop.flink.model.impl.operators.matching.preserving.explorative.debug;
 
+import org.apache.commons.lang3.StringUtils;
 import org.apache.log4j.Logger;
 import org.gradoop.flink.model.impl.operators.matching.common.debug.Printer;
-import org.gradoop.flink.model.impl.operators.matching.isomorphism
-  .explorative.tuples.VertexStep;
+import org.gradoop.flink.model.impl.operators.matching.preserving.explorative.tuples.EmbeddingWithTiePoint;
+
+
+
+import java.util.Arrays;
 
 /**
- * Debug output for {@link VertexStep}.
+ * Debug output for {@link EmbeddingWithTiePoint}.
  *
  * @param <K> key type
  */
-public class PrintVertexStep<K>
-  extends Printer<VertexStep<K>, K> {
+public class PrintEmbeddingWithTiePoint<K>
+  extends Printer<EmbeddingWithTiePoint<K>, K> {
   /**
    * Logger
    */
-  private static Logger LOG = Logger.getLogger(PrintVertexStep.class);
+  private static final Logger LOG = Logger.getLogger(
+    PrintEmbeddingWithTiePoint.class);
+
+  /**
+   * Constructor
+   */
+  public PrintEmbeddingWithTiePoint() {
+    this(false, "");
+  }
 
   /**
    * Constructor
@@ -40,13 +52,18 @@ public class PrintVertexStep<K>
    * @param isIterative true, if used in iterative context
    * @param prefix      prefix for debug string
    */
-  public PrintVertexStep(boolean isIterative, String prefix) {
+  public PrintEmbeddingWithTiePoint(boolean isIterative, String prefix) {
     super(isIterative, prefix);
   }
 
   @Override
-  protected String getDebugString(VertexStep<K> vertexStep) {
-    return String.format("(%s)", vertexMap.get(vertexStep.getVertexId()));
+  protected String getDebugString(EmbeddingWithTiePoint<K> embedding) {
+    return String.format("(([%s],[%s]),%s)",
+      StringUtils.join(convertList(Arrays.asList(
+        embedding.getEmbedding().getVertexMappings()), true), ','),
+      StringUtils.join(convertList(Arrays.asList(
+        embedding.getEmbedding().getEdgeMappings()), false), ','),
+      vertexMap.get(embedding.getTiePointId()));
   }
 
   @Override
