+++ gradoop-flink/src/main/java/org/gradoop/flink/model/impl/operators/matching/preserving/explorative/functions/VertexHasCandidate.java
@@ -15,16 +15,15 @@
  * along with Gradoop. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.gradoop.flink.model.impl.operators.matching.isomorphism.explorative.functions;
+package org.gradoop.flink.model.impl.operators.matching.preserving.explorative.functions;
+
 
 import org.apache.flink.api.common.functions.RichFilterFunction;
 import org.apache.flink.api.java.functions.FunctionAnnotation;
 import org.apache.flink.configuration.Configuration;
-import org.gradoop.flink.model.impl.operators.matching.common.tuples
-  .IdWithCandidates;
 import org.gradoop.flink.model.impl.operators.matching.common.query.TraversalCode;
-
-import org.gradoop.flink.model.impl.operators.matching.isomorphism.explorative.ExplorativeSubgraphIsomorphism;
+import org.gradoop.flink.model.impl.operators.matching.common.tuples.IdWithCandidates;
+import org.gradoop.flink.model.impl.operators.matching.preserving.explorative.ExplorativePatternMatching;
 
 /**
  * Filters vertices if their candidates contain a given candidate. This
@@ -64,7 +63,7 @@ public class VertexHasCandidate<K>
   public void open(Configuration parameters) throws Exception {
     super.open(parameters);
     int step = (int) getRuntimeContext()
-      .getBroadcastVariable(ExplorativeSubgraphIsomorphism.BC_SUPERSTEP).get(0);
+      .getBroadcastVariable(ExplorativePatternMatching.BC_SUPERSTEP).get(0);
     candidate = (int) traversalCode.getStep(step - 1).getTo();
   }
