+++ gradoop-flink/src/main/java/org/gradoop/flink/model/impl/operators/matching/preserving/package-info.java
@@ -15,35 +15,8 @@
  * along with Gradoop. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package org.gradoop.flink.model.impl.operators.matching.isomorphism.explorative.tuples;
-
-import org.apache.flink.api.java.tuple.Tuple2;
-import org.gradoop.flink.model.impl.operators.matching.common.tuples.Embedding;
-
 /**
- * Represents an embedding and a weld point to grow the embedding.
- *
- * f0: tie point
- * f1: embedding
- *
- * @param <K> key type
+ * Contains structure-preserving pattern matching algorithms, i.e., such based
+ * on homomorphisms and isomorphisms between input and pattern graph.
  */
-public class EmbeddingWithTiePoint<K>
-  extends Tuple2<K, Embedding<K>> {
-
-  public K getTiePointId() {
-    return f0;
-  }
-
-  public void setTiePointId(K tiePointId) {
-    f0 = tiePointId;
-  }
-
-  public Embedding<K> getEmbedding() {
-    return f1;
-  }
-
-  public void setEmbedding(Embedding<K> embedding) {
-    f1 = embedding;
-  }
-}
+package org.gradoop.flink.model.impl.operators.matching.preserving;
